requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0  # Changed from anthropic to gemini

# Data processing
python-dateutil>=2.8.0
feedparser>=6.0.0  # RSS feed parsing for CoinDesk/CoinTelegraph

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.0.0
flake8>=6.0.0

# Optional
praw>=7.7.0  # Reddit API
python-dotenv>=1.0.0  # Environment variables
